refactor: #13 architecture
クラス分離
