+++ .gitignore
@@ -5,3 +5,4 @@ config.json
 dist
 build
 **/*.spec
+**/__pycache__
